+++ images/mcp_fuzzer_arch.png
