+++ mcp_fuzzer/__init__.py
@@ -1,0 +1,16 @@
+"""
+MCP Fuzzer - A CLI tool for fuzzing MCP server tools using multiple transport protocols.
+"""
+
+from .client import fuzz_tool, main
+from .transport import create_transport, TransportProtocol
+from .strategies import make_fuzz_strategy_from_jsonschema
+
+__version__ = "0.1.0"
+__all__ = [
+    "fuzz_tool",
+    "main", 
+    "create_transport",
+    "TransportProtocol",
+    "make_fuzz_strategy_from_jsonschema"
+]
